backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.28.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-multipart==0.0.9
google-genai==1.5.0
